update users controllers, add gwtuserbyid middleware
